Refactored error handling in Coral.Managed to utilize the message callback + added automatic object cleanup on assembly unload
- Changed StableVector m_PageCount to uint64_t (caused problems with std::max on some compilers)

+++ Coral.Native/Include/Coral/StableVector.hpp
@@ -209,7 +209,7 @@ namespace Coral {
 		std::atomic<Page**> m_PageTable;
 		std::atomic<uint32_t> m_ElementCount = 0;
 		std::atomic<uint32_t> m_Capacity = 0;
-		size_t m_PageCount = 0;
+		uint64_t m_PageCount = 0;
 	};
 
 }
